agents/cv-agent/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cv-agent"
version = "0.1.0"
description = "Agent that tailors CVs to job descriptions"
authors = [{name = "lifesheet", email = "devnull@example.com"}]
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.111",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.7",
  "python-dotenv>=1.0",
]

[project.optional-dependencies]
ai = [
  # Google's Generative AI SDK
  "google-generativeai>=0.5; python_version>='3.10'",
]
dev = [
  "pytest>=8.0",
  "httpx>=0.27",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-q"
pythonpath = ["src"]

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
